% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{altDel}
\alias{altDel}
\title{CREATE ALTERNATIVE SEQUENCES FOR DELETIONS}
\usage{
altDel(reference_sequence, deletion_bases)
}
\arguments{
\item{145bp}{reference sequence, string containing bases to delete.}
}
\value{
alt sequence with specified bases removed (counts the number of bases
in string, removes bases starting at 73)
}
\description{
Delete specified number of bases from ref seq. (Counts the number of bases
in deletion string, removes bases starting at 73)
}
