% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{altIns}
\alias{altIns}
\title{CREATE ALTERNATIVE SEQUENCES FOR INSERTIONS}
\usage{
altIns(reference_sequence, insertion_bases)
}
\arguments{
\item{145bp}{reference sequence, string containing bases to insert.}
}
\value{
145bp alt sequence with specified bases inserted
}
\description{
Insert specified bases after base 73 into provided ref seq. Shortens sequence
from both sides so total length is <= 145bp.
}
