% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{altSnps}
\alias{altSnps}
\title{CREATE ALTERNATIVE SEQUENCES FOR SNPS}
\usage{
altSnps(reference_sequence, new_base)
}
\arguments{
\item{145bp}{reference sequence, string containing base to swap in.}
}
\value{
alt sequence with specified base swapped in
}
\description{
Change base 73 to alt base.
}
