% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{delBases}
\alias{delBases}
\title{GET BASES TO BE DELETED}
\usage{
delBases(reference_sequence, deletion_bases)
}
\arguments{
\item{reference}{sequence, string containing bases to delete}
}
\value{
string containing bases that the program will delete.
}
\description{
Check which bases the program will delete. Can be used to compare to
the bases that should be deleted per variant input file.
}
