% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fixDigWild}
\alias{fixDigWild}
\title{FIX WildType DIGESTION SITE}
\usage{
fixDigWild(enzx, enzFIX, sequence, coord1)
}
\arguments{
\item{145bp}{sequence containing site, coordinate of first base of
restriction site, coordinate of second base of restriction site.}
}
\value{
corrected sequence
}
\description{
Changes SfiI site (GGCCNNNNNGGCC -> GCGCNNNNNGGCC)
Will only first first digestion site in sequence.
}
