% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{generate}
\alias{generate}
\title{Generate Design File}
\usage{
generate(variant_input_path, tag_path, scrambled_path)
}
\arguments{
\item{variant}{input path, tags path, scrambled variants path}
}
\value{
design file
}
\description{
To actually generate the file containing all oligos in the library, run this function.
}
