% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{refIns}
\alias{refIns}
\title{SHORTEN REFERENCE SEQUENCES FOR INSERTIONS}
\usage{
refIns(reference_sequence, insertion_bases)
}
\arguments{
\item{145bp}{reference sequence, string containing bases that were inserted.}
}
\value{
ref sequence with appropriate end bases removed
}
\description{
Ensures ref seq for insertions has the same bases removed
from the ends as alt seq.
}
