% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{snpBases}
\alias{snpBases}
\title{GET BASE TO BE CHANGED}
\usage{
snpBases(reference_sequence)
}
\arguments{
\item{reference}{sequence}
}
\value{
string containing base that the program will change.
}
\description{
Check which base the program will change. Can be used to compare to
the bases that should be changed per variant input file.
}
