% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{generate}
\alias{generate}
\title{Generate Design File}
\usage{
generate(tags_per_variant, variant_input_path, tag_path, scrambled_path)
}
\arguments{
\item{tags_per_variant}{number of oligos with unique tags created for each variant}

\item{variant_input_path}{path to file with variant input (e.g. "Documents/input_files/variant_input.csv")}

\item{tag_path}{path to file containing tags}

\item{scrambled_path}{path to file containing scrambled sequences}
}
\value{
design file
}
\description{
To actually generate the file containing all oligos in the library, run this function.
}
